Move the package filtering responsibility to getImportablePackages
